--- backend/config.js
@@ -1,3 +1,0 @@
-export default {
-    baseUrl: process.env.BASE_URL || "/",
-};
