+++ backend/routes/index.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import eventRoutes from './eventRoutes.js';
+
+const router = express.Router();
+
+router.use('/api', eventRoutes);
+
+export default router;
